fix: keep files route exports build-safe
